Moving oscan1_ftdi.c logic into ftdi.c.
