+++ tcl/interface/ftdi/olimex-arm-jtag-cjtag.cfg
@@ -9,18 +9,19 @@
 # http://www.olimex.com/dev/arm-usb-tiny-h.html
 #
 
-interface oscan1_ftdi
-oscan1_ftdi_device_desc "Olimex OpenOCD JTAG ARM-USB-TINY-H"
-oscan1_ftdi_vid_pid 0x15ba 0x002a
+interface ftdi
+ftdi_oscan1_mode on
+ftdi_device_desc "Olimex OpenOCD JTAG ARM-USB-TINY-H"
+ftdi_vid_pid 0x15ba 0x002a
 
-oscan1_ftdi_layout_init 0x0808 0x0a1b
-oscan1_ftdi_layout_signal nSRST -oe 0x0200
+ftdi_layout_init 0x0808 0x0a1b
+ftdi_layout_signal nSRST -oe 0x0200
 # oscan1_ftdi_layout_signal nTRST -data 0x0100 -oe 0x0100
-oscan1_ftdi_layout_signal LED -data 0x0800
+ftdi_layout_signal LED -data 0x0800
 
 # These signals are used for cJTAG escape sequence on initialization only
-oscan1_ftdi_layout_signal TCK -data 0x0001
-oscan1_ftdi_layout_signal TDI -data 0x0002
-oscan1_ftdi_layout_signal TDO -input 0x0004
-oscan1_ftdi_layout_signal TMS -data 0x0008
-oscan1_ftdi_layout_signal JTAG_SEL -data 0x0100 -oe 0x0100
+ftdi_layout_signal TCK -data 0x0001
+ftdi_layout_signal TDI -data 0x0002
+ftdi_layout_signal TDO -input 0x0004
+ftdi_layout_signal TMS -data 0x0008
+ftdi_layout_signal JTAG_SEL -data 0x0100 -oe 0x0100
